% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileDist-class.R
\name{backgroundDist}
\alias{backgroundDist}
\title{Background distribution.}
\usage{
backgroundDist(x)
}
\arguments{
\item{x}{A ProfileDist object.}
}
\value{
The background distribution vector.
}
\description{
\code{backgroundDist} returns the background distribution of a profile
object.
}
\details{
This is a generic function.
}
\examples{
anObject <- ProfileDist()
backgroundDist(anObject)
}

