% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INR}
\alias{INR}
\title{The position count matrix for INR.}
\format{A matrix with named rows corresponding to the counts for each of
  the four nucleotides.}
\source{
http://jaspar.genereg.net
}
\usage{
INR
}
\value{
The position count matrix for INR.
}
\description{
The position count matrix for the initiator (INR) core promoter element. This
matrix was obtained from the JASPAR public domain database, but was
originally published by P. Bucher (1990); in that publication (and elsewhere)
it was termed Cap signal.
}
\references{
Bucher, P. (1990). Weight matrix descriptions of four eukaryotic
  RNA polymerase II promoter elements derived from 502 unrelated promoter
  sequences. Journal of Molecular Biology 212, 563???-578.

Mathelier, A., Zhao, X., Zhang, A.W., Parcy, F., Worsley-Hunt,
  R., Arenillas, D.J., Buchman, S., Chen, C.-Y., Chou, A., Ienasescu, H., et
  al. (2014). JASPAR 2014: an extensively expanded and updated open-access
  database of transcription factor binding profiles. Nucleic Acids Research
  42, D142--D147.
}
\keyword{datasets}

