% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileDist-class.R
\name{plotDist}
\alias{plotDist}
\title{Plot background and signal distributions.}
\usage{
plotDist(x)
}
\arguments{
\item{x}{A ProfileDist object.}
}
\value{
The scores vector.
}
\description{
\code{plotDist} creates a rudimentary plot of signals and backgrounds.
}
\details{
This is a generic function.
}
\examples{
data(INR)
thedist <- computeScoreDist(regularizeMatrix(INR), 0.5)
plotDist(thedist)
}

