% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileDist-class.R
\docType{class}
\name{ProfileDist-class}
\alias{ProfileDist}
\alias{ProfileDist-class}
\alias{backgroundDist,ProfileDist-method}
\alias{plotDist,ProfileDist-method}
\alias{score,ProfileDist-method}
\alias{show,ProfileDist-method}
\alias{signalDist,ProfileDist-method}
\title{ProfileDist}
\usage{
\S4method{show}{ProfileDist}(object)

\S4method{score}{ProfileDist}(x)

\S4method{signalDist}{ProfileDist}(x)

\S4method{backgroundDist}{ProfileDist}(x)

\S4method{plotDist}{ProfileDist}(x)
}
\arguments{
\item{object}{A ProfileDist object for the \code{show} method.}

\item{x}{A ProfileDist object.}
}
\value{
A ProfileDist object.
}
\description{
This class represents signal and background score distributions for a
profile.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows useful information

\item \code{score}: Accessor for the scores

\item \code{signalDist}: Accessor for the signal distribution

\item \code{backgroundDist}: Accessor for the background distribution

\item \code{plotDist}: Simple plot method for signal and background
distributions
}}
\section{Slots}{

\describe{
\item{\code{f}}{Signal distribution}

\item{\code{g}}{Background distribution}

\item{\code{Scores}}{Scores for the distributions}
}}
\section{Constructor}{

\code{ProfileDist(f=numeric, g=numeric, Scores=numeric)}
}

