% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileDist-class.R
\name{signalDist}
\alias{signalDist}
\title{Signal distribution.}
\usage{
signalDist(x)
}
\arguments{
\item{x}{A ProfileDist object.}
}
\value{
The signal distribution vector.
}
\description{
\code{signalDist} returns the signal distribution of a profile
object.
}
\details{
This is a generic function.
}
\examples{
anObject <- ProfileDist()
backgroundDist(anObject)
}

